third_party/blink/web_tests/http/tests/security/sandboxed-iframe-form-top-expected.txt

CONSOLE ERROR: line 8: Unsafe JavaScript attempt to initiate navigation for frame with URL 'http://127.0.0.1:8000/security/sandboxed-iframe-form-top.html' from frame with URL 'http://127.0.0.1:8000/security/resources/sandboxed-iframe-form-top.html'. The frame attempting navigation of the top-level window is sandboxed, but the flag of 'allow-top-navigation' or 'allow-top-navigation-by-user-activation' is not set.

CONSOLE ERROR: line 8: Blocked opening 'http://127.0.0.1:8000/security/resources/fail.html?' in a new window because the request was made in a sandboxed frame whose 'allow-popups' permission is not set.
This tests passes if the sandboxed frame cannot navigate the top frame.

PASS
